supplement of function (#52)
* test expressions,operators,list,map

* test patterns

* test temporal value

* format

* test spatial values

* format class name

* add comments

* change cypher format

* implement percentitleCont function

* reverse function syns

* implement reverse,tail function

* implement stDev function

* implement stDev,stDevP function

* implement stDev,stDevP,percentileDesc

* change scala version

* change searchNearestNum to upBound

* change percentileDisc function return type

* implement randomUUID function

* repair bug of collect function

* change data type in test of usingCountTotoGroupAndCountRelationshipTypes()

* repair some test bugs of E_Mathematical_Numeric.scala

* repair data type bug in F_Mathematical_Logarithmic

* implement haversin() function

* change data type of operand in Multiply

+++ src/test/main/scala/org/grapheco/cypher/functions/F_Mathematical_Logarithmic.scala
@@ -71,7 +71,7 @@ class F_Mathematical_Logarithmic extends TestBase {
         |""".stripMargin).records().toArray
 
     Assert.assertEquals(1, records.length)
-    Assert.assertEquals(16.toFloat, records(0)("sqrt(256)").asInstanceOf[LynxValue].value)
+    Assert.assertEquals(16.0, records(0)("sqrt(256)").asInstanceOf[LynxValue].value)
   }
